+++ packages/common/src/services/__tests__/grid.service.spec.ts
@@ -1,5 +1,3 @@
-import 'jest-extended';
-
 import { FilterService, GridService, GridStateService, PaginationService, PubSubService, SharedService, SortService, TreeDataService } from '../index';
 import { GridOption, CellArgs, Column, OnEventArgs, SlickGrid, SlickDataView, SlickNamespace } from '../../interfaces/index';
 import { SlickRowSelectionModel } from '../../plugins/slickRowSelectionModel';
@@ -20,9 +18,13 @@ const mockRowSelectionModel = {
   onSelectedRangesChanged: new Slick.Event(),
 } as unknown as SlickRowSelectionModel;
 
+<<<<<<< HEAD
 jest.mock('../../plugins/slickRowSelectionModel', () => ({
   SlickRowSelectionModel: jest.fn().mockImplementation(() => mockRowSelectionModel),
 }));
+=======
+jest.mock('flatpickr', () => { });
+>>>>>>> master
 
 const filterServiceStub = {
   clearFilters: jest.fn(),
@@ -102,6 +104,9 @@ const treeDataServiceStub = {
 } as unknown as TreeDataService;
 
 describe('Grid Service', () => {
+  jest.mock('slickgrid/plugins/slick.rowselectionmodel', () => mockSelectionModelImplementation);
+  Slick.RowSelectionModel = mockSelectionModelImplementation;
+
   let service: GridService;
   const sharedService = new SharedService();
   const mockGridOptions = { enableAutoResize: true } as GridOption;
@@ -1137,6 +1142,26 @@ describe('Grid Service', () => {
 
       expect(setColumnsSpy).toHaveBeenCalled();
       expect(setOptionsSpy).toHaveBeenCalledWith({ frozenBottom: false, frozenColumn: -1, frozenRow: -1, enableMouseWheelScrollHandler: false });
+    });
+
+    it('should call "setPinning" which itself calls "clearPinning" when the pinning option input is an empty object', () => {
+      const mockPinning = {};
+      const clearPinningSpy = jest.spyOn(service, 'clearPinning');
+      jest.spyOn(SharedService.prototype, 'slickGrid', 'get').mockReturnValue(gridStub);
+
+      service.setPinning(mockPinning);
+
+      expect(clearPinningSpy).toHaveBeenCalled();
+    });
+
+    it('should call "setPinning" which itself calls "clearPinning" when the pinning option input is null', () => {
+      const mockPinning = null;
+      const clearPinningSpy = jest.spyOn(service, 'clearPinning');
+      jest.spyOn(SharedService.prototype, 'slickGrid', 'get').mockReturnValue(gridStub);
+
+      service.setPinning(mockPinning);
+
+      expect(clearPinningSpy).toHaveBeenCalled();
     });
 
     it('should call "setPinning" and expect SlickGrid "setOptions" be called with new frozen options and "autosizeColumns" also be called', () => {