+++ packages/common/src/services/__tests__/sort.service.spec.ts
@@ -1157,6 +1157,19 @@ describe('SortService', () => {
         sharedService.hierarchicalDataset = dataset;
       });
 
+      it('should sort the hierarchical dataset and expect event emitted when passing True as 3rd argument', () => {
+        const pubSubSpy = jest.spyOn(pubSubServiceStub, 'publish');
+        const sortTreeDataSpy = jest.spyOn(service, 'sortTreeData');
+        const emitSortChangedSpy = jest.spyOn(service, 'emitSortChanged');
+
+        const result = service.sortHierarchicalDataset(dataset, [{ columnId: 'file', sortAsc: true, sortCol: mockColumns[0] }], true);
+
+        expect(result).toBeTruthy();
+        expect(pubSubSpy).toHaveBeenCalledWith('onSortChanged', [{ columnId: 'file', direction: 'ASC' }]);
+        expect(sortTreeDataSpy).toHaveBeenCalled();
+        expect(emitSortChangedSpy).toHaveBeenCalled();
+      });
+
       it('should call onLocalSortChanged with a hierarchical dataset and expect DataView "setItems" method be called once with sorted ASC dataset', (done) => {
         gridOptionMock.enableTreeData = true;
         gridOptionMock.treeDataOptions = { columnId: 'file', childrenPropName: 'files', };