+++ packages/vanilla-bundle/src/services/index.ts
@@ -1,2 +1,1 @@
-export * from './textExport.service';
 export * from './universalContainer.service';